+++ Source/DynamicScene/StaticGeometryColorBatch.js
@@ -50,7 +50,7 @@ define(['../Core/Color',
     };
 
     Batch.prototype.update = function(time) {
-        var canAnimate = true;
+        var isUpdated = true;
         var removedCount = 0;
         var primitive = this.primitive;
         var primitives = this.primitives;
@@ -71,7 +71,7 @@ define(['../Core/Color',
                 });
 
                 primitives.add(primitive);
-                canAnimate = false;
+                isUpdated = false;
             }
             this.primitive = primitive;
             this.createPrimitive = false;
@@ -102,10 +102,10 @@ define(['../Core/Color',
                 }
             }
         } else if (defined(primitive) && primitive._state !== PrimitiveState.COMPLETE) {
-            canAnimate = false;
+            isUpdated = false;
         }
         this.itemsToRemove.length = removedCount;
-        return canAnimate;
+        return isUpdated;
     };
 
     Batch.prototype.removeAllPrimitives = function() {
@@ -146,8 +146,8 @@ define(['../Core/Color',
         var updater;
 
         //Perform initial update
-        var canAnimate = this._solidBatch.update(time);
-        canAnimate = this._translucentBatch.update(time) && canAnimate;
+        var isUpdated = this._solidBatch.update(time);
+        isUpdated = this._translucentBatch.update(time) && isUpdated;
 
         //If any items swapped between solid/translucent, we need to
         //move them between batches
@@ -173,11 +173,11 @@ define(['../Core/Color',
 
         //If we moved anything around, we need to re-build the primitive
         if (solidsToMoveLength > 0 || translucentToMoveLength > 0) {
-            canAnimate = this._solidBatch.update(time) && canAnimate;
-            canAnimate = this._translucentBatch.update(time) && canAnimate;
+            isUpdated = this._solidBatch.update(time) && isUpdated;
+            isUpdated = this._translucentBatch.update(time) && isUpdated;
         }
 
-        return canAnimate;
+        return isUpdated;
     };
 
     StaticGeometryColorBatch.prototype.removeAllPrimitives = function() {